remove IsolatedStorageStrategy
