+++ .gitignore
@@ -15,3 +15,4 @@ AjaxControlToolkit.Reference/bin/
 .vs/
 ReferenceCore/bin/
 AjaxControlToolkit.SampleSite/vwd.webinfo
+AjaxControlToolkit.Jasmine/_AjaxFileUpload/*
